+++ Cargo.toml
@@ -9,8 +9,10 @@ edition = "2018"
 log = "0.4.8"
 serde_json = "1.0.41"
 serde_yaml = "0.8.11"
-mongodb = "0.4.0"
-bson = "0.14.0"
+mongodb = "0.3.12"
+bson = "0.13.0"
+#mongodb = "0.4.0"
+#bson = "0.14.0"
 
 data-finder = { git = "https://github.com/pewpew79/data-finder.git", branch = "master" }
 data-getter = { git = "https://github.com/pewpew79/data-getter.git", branch = "master" }
