--- src_new/core/mod.rs
@@ -1,6 +1,0 @@
-pub mod functions;
-pub mod io_utils;
-pub mod data_utils;
-pub mod common_utils;
-pub mod config_utils;
-pub mod storage_utils;
